Created Test package and class

+++ JavaProject/src/Facade/AdminFacade.java
@@ -2,7 +2,6 @@ package Facade;
 
 import java.util.ArrayList;
 
-import DBDAO.CompaniesDBDAO;
 import JavaBeans.Company;
 import JavaBeans.Coupon;
 import JavaBeans.Customer;
@@ -19,7 +18,7 @@ public class AdminFacade extends ClientFacade {
 		//-----------------------Methods ---------------------------
 
 		public boolean login(String email,String password) {
-			if (email == "admin@admin.com" && password == "admin")
+			if (email.contentEquals("admin@admin.com") && password.contentEquals("admin"))
 				return true;
 			else
 				return false;
@@ -31,33 +30,47 @@ public class AdminFacade extends ClientFacade {
 			ArrayList<Company> allCompanies = companiesDAO.getAllCompanies();
 			for (Company c : allCompanies)
 				//If a Company with the same Email or Name as the new Company exists, print an appropriate message and quit the function
-				if (c.getName() == company.getName() || c.getEmail() == company.getEmail()) {
+				if (c.getName().contentEquals(company.getName()) || c.getEmail().contentEquals(company.getEmail())) {
 					System.out.println("A company with the same Name/Email already exists");
 					return;
 				}
 					
 			//Adding the Company after validating it doesn't have an identical Name or Email
 			companiesDAO.addCompany(company);
-			System.out.println("Company "+company.getName()+" added");
 		}
 		
 		public void updateCompany(Company company) throws Exception {
 			Company theCompany = companiesDAO.getOneCompany(company.getId());
-			theCompany.setEmail(company.getEmail());
-			theCompany.setPassword(company.getPassword());
-			theCompany.setCoupons(company.getCoupons());
-			companiesDAO.updateCompany(theCompany);
+			
+			if (theCompany != null) {
+				theCompany.setEmail(company.getEmail());
+				theCompany.setPassword(company.getPassword());
+				theCompany.setCoupons(company.getCoupons());
+				companiesDAO.updateCompany(theCompany);
+			}
 		}
 		
 		public void deleteCompany(int companyID) throws Exception {
 			
+			//Getting a list of All the customers in order to delete all the coupon purchases by this company
+			ArrayList<Customer> allCustomers = customersDAO.getAllCustomers();
+
 			//Getting the coupons that the company has
 			Company theCompany = companiesDAO.getOneCompany(companyID);
-			ArrayList<Coupon> theCoupons = theCompany.getCoupons();
-			
-			//Deleting the Coupons that the Company created
-			for (Coupon c : theCoupons)
-				couponsDAO.deleteCoupon(c.getId());
+			if (theCompany != null) {
+				ArrayList<Coupon> theCoupons = theCompany.getCoupons();
+				
+				//Deleting the Coupons that the Company created and the history of Coupon Purchases by this company
+				for (Coupon theCoupon : theCoupons) {
+					//Deleting the history of Coupon Purchases by this company
+					for (Customer theCustomer: allCustomers) {
+						couponsDAO.deleteCouponPurchase(theCustomer.getId(),theCoupon.getId());
+					}
+					
+					//Deleting the Coupon that this created
+					couponsDAO.deleteCoupon(theCoupon.getId());
+				}
+			}
 			
 			
 			//Deleting the company itself
@@ -75,45 +88,52 @@ public class AdminFacade extends ClientFacade {
 		
 		public void addCustomer(Customer customer) throws Exception {
 			ArrayList<Customer> allCustomers = customersDAO.getAllCustomers();
-			for (Customer c : allCustomers)
-				if (customer.getEmail() == c.getEmail()) {
+			for (Customer c : allCustomers) {
+				if (customer.getEmail().contentEquals(c.getEmail())) {
 					System.out.println("A Customer with the same Email already exists");
 					return;
 				}
+			}
 			
 			customersDAO.addCustomer(customer);
-			System.out.println("Customer "+customer.getFirstName()+" "+customer.getLastName()+" added");
 		}
 		
 		public void updateCustomer(Customer customer) throws Exception {
 			
-			//Creating a new Customer with the details that are about to be updated (all the details but the ID)
+			//Getting a Customer with the details that are about to be updated (all the details but the ID)
 			Customer theCustomer = customersDAO.getOneCustomer(customer.getId());
-			theCustomer.setEmail(customer.getEmail());
-			theCustomer.setFirstName(customer.getFirstName());
-			theCustomer.setLastName(customer.getLastName());
-			theCustomer.setPassword(customer.getPassword());
-			theCustomer.setCoupons(customer.getCoupons());
 			
-			//Updating the Customer
-			customersDAO.updateCustomer(theCustomer);
-
+			if (theCustomer != null) {
+				theCustomer.setEmail(customer.getEmail());
+				theCustomer.setFirstName(customer.getFirstName());
+				theCustomer.setLastName(customer.getLastName());
+				theCustomer.setPassword(customer.getPassword());
+				theCustomer.setCoupons(customer.getCoupons());
+				
+				//Updating the Customer
+				customersDAO.updateCustomer(theCustomer);
+			}
 		}
 		
 		public void deleteCustomer(int customerID) throws Exception {
 			
-			//Deleting all of the coupons that were created by the specific Customer
-			
-			//getting all coupons
+			//Getting the Customer's details from the Database
 			Customer theCustomer = customersDAO.getOneCustomer(customerID);
-			ArrayList<Coupon> theCoupons = theCustomer.getCoupons();
 			
-			//all coupons that include this specific customerID will be deleted
-			for (Coupon c : theCoupons)
-				couponsDAO.deleteCouponPurchase(customerID,c.getId());
-			
-			//Delete the Customer
-			customersDAO.deleteCustomer(customerID);
+			//Delete the customer only if he exists
+			if (theCustomer != null) {		
+				
+				//Getting the Customer's Coupons
+				ArrayList<Coupon> theCoupons = theCustomer.getCoupons();
+				
+				//all coupon purchases that include this specific customerID will be deleted
+				for (Coupon c : theCoupons) {
+					couponsDAO.deleteCouponPurchase(customerID,c.getId());
+				}
+				
+				//Delete the Customer
+				customersDAO.deleteCustomer(customerID);
+			}
 		}
 		
 		public ArrayList<Customer> getAllCustomers() throws Exception{
